# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST de Gestión de Vehículos",
        "description": "API RESTful para gestión de vehículos, marcas y propietarios desarrollada con Laravel",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "contacto@vehiculos-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/api/marcas-vehiculo": {
            "get": {
                "tags": [
                    "Marcas de Vehículos"
                ],
                "summary": "Obtener lista de todas las marcas de vehículos",
                "description": "Retorna una lista completa de todas las marcas de vehículos registradas",
                "operationId": "e7e54245987f74a1dd937cec842771c8",
                "responses": {
                    "200": {
                        "description": "Lista de marcas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marcas obtenidas exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VehicleBrand"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Marcas de Vehículos"
                ],
                "summary": "Crear una nueva marca de vehículo",
                "description": "Registra una nueva marca de vehículo en el sistema",
                "operationId": "ad21c70580cccc2127bf95d5db96c291",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_name",
                                    "country"
                                ],
                                "properties": {
                                    "brand_name": {
                                        "description": "Nombre de la marca",
                                        "type": "string",
                                        "example": "Tesla"
                                    },
                                    "country": {
                                        "description": "País de origen",
                                        "type": "string",
                                        "example": "Estados Unidos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marca creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marca de vehículo creada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VehicleBrand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/marcas-vehiculo/{id}": {
            "get": {
                "tags": [
                    "Marcas de Vehículos"
                ],
                "summary": "Obtener una marca específica",
                "description": "Retorna los detalles de una marca de vehículo por su ID",
                "operationId": "ea4e7e2a18dc4bbed82528b68f6b9ad0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marca obtenida exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VehicleBrand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Marcas de Vehículos"
                ],
                "summary": "Actualizar una marca existente",
                "description": "Actualiza los datos de una marca de vehículo",
                "operationId": "e563bdb65b8833e2aabb94529060f7b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand_name",
                                    "country"
                                ],
                                "properties": {
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Tesla Motors"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Estados Unidos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marca actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marca de vehículo actualizada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VehicleBrand"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Marcas de Vehículos"
                ],
                "summary": "Eliminar una marca",
                "description": "Elimina una marca de vehículo del sistema (también elimina vehículos asociados en cascada)",
                "operationId": "cefa1f79fe0e5b094c83ffd7a02a53f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marca de vehículo eliminada exitosamente"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error en la operación"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error de validación"
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "El campo es obligatorio."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "VehicleBrand": {
                "title": "Marca de Vehículo",
                "description": "Modelo de marca de vehículo",
                "properties": {
                    "id": {
                        "description": "ID de la marca",
                        "type": "integer",
                        "example": 1
                    },
                    "brand_name": {
                        "description": "Nombre de la marca",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "country": {
                        "description": "País de origen",
                        "type": "string",
                        "example": "Japón"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10 10:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10 10:30:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}